Move problem type subclasses into separate files
